feature / meeyoung / #17 물품 상세페이지 이미지 불러오는 부분 컴포넌트로 분리
