+++ shoppingmall/src/components/productDetails/ProductDetail.js
@@ -3,11 +3,14 @@ import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 
 // svg
-import { FiCircle } from "react-icons/fi";
+// import { FiCircle } from "react-icons/fi";
 import leftarrow from "../../assets/leftarrow.svg";
 import profileimage from "../../assets/profileimage.svg";
 import unlike from "../../assets/unlike.svg";
-import rightarrow from "../../assets/rightarrow.svg";
+// import rightarrow from "../../assets/rightarrow.svg";
+
+// component
+import ProductCarousel from './ProductCarousel';
 
 
 const ProductDetail = () => {
@@ -24,16 +27,7 @@ const ProductDetail = () => {
                     <BackIcon src={leftarrow} onClick={navigateToPage} />
                     <ProductTitle>상품 정보</ProductTitle>
                 </Header>
-                <ImageWrapper>
-                    <ProductImage src="https://via.placeholder.com/400/#D9D9D9"/>
-                    <LeftArrow src={leftarrow}/> 
-                    <RightArrow src={rightarrow}/>
-                    <CarouselButtons>
-                        <PaginationButton />
-                        <PaginationButton />
-                        <PaginationButton />
-                    </CarouselButtons>
-                </ImageWrapper>
+                <ProductCarousel/>
                 <ContentWrapper>
                     <UserInfo>
                         <UserImage src={profileimage} alt="personIcon"/>
@@ -100,47 +94,6 @@ const ProductTitle = styled.div`
     font-weight: bold;
     font-size: 24px;
 `;
-
-const ImageWrapper = styled.div`
-    display: flex;
-    justify-content:center;
-    width: 700px;
-    hegith: 448px;
-    margin-top: 55px;
-`;
-
- const ProductImage = styled.img`
-    width: 400px;
-    height: 400px;
-    border-radius: 10px;
-`;
-
-const LeftArrow = styled.img`
-    position: absolute;
-    top: 410px;
-    left: 480px;
-    cursor: pointer;
-`;
-
-const RightArrow = styled.img`
-    position: absolute;
-    top: 410px;
-    right: 444px;
-    cursor: pointer;
-`;
-
-const CarouselButtons = styled.div`
-    position: absolute;
-    cursor: pointer;
-    top: 580px;
-    right: 600px;
-`;
-
-const PaginationButton = styled(FiCircle)`
-    margin-right: 4px;
-    background-color: none;
-`;
-
 
 const ContentWrapper = styled.div`
     display: flex;
